+++ GMap.NET.Core/GMap.NET.MapProviders/CzechHistoryMapProvider.cs
@@ -65,8 +65,8 @@ namespace GMap.NET.MapProviders
       {
          // http://m4.mapserver.mapy.cz/army2/9_7d00000_8080000
 
-         int xx = pos.X << (28 - zoom);
-         int yy = ((((int)Math.Pow(2.0, (double)zoom)) - 1) - pos.Y) << (28 - zoom);
+         long xx = pos.X << (28 - zoom);
+         long yy = ((((long)Math.Pow(2.0, (double)zoom)) - 1) - pos.Y) << (28 - zoom);
 
          return string.Format(UrlFormat, GetServerNum(pos, 3) + 1, zoom, xx, yy);
       }
